+++ light-client/src/utils/block_on.rs
@@ -11,11 +11,10 @@ where
     F::Output: Send,
 {
     std::thread::spawn(move || {
-        let mut rt = tokio::runtime::Builder::new()
-            .basic_scheduler()
+        let rt = tokio::runtime::Builder::new_current_thread()
             .enable_all()
             .build()
-            .unwrap();
+            .map_err(|_| IoError::Runtime)?;
 
         if let Some(timeout) = timeout {
             let task = async { tokio::time::timeout(timeout, f).await };